requirements.txt
asttokens==3.0.0
comm==0.2.3
debugpy==1.8.15
decorator==5.2.1
executing==2.2.0
iniconfig==2.3.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyMuPDF==1.26.3
pytesseract==0.3.13
pytest==9.0.2
python-dateutil==2.9.0.post0
pyzmq==27.0.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
wcwidth==0.2.13
opencv-python==4.10.0.84
numpy==1.26.4
scikit-image==0.24.0
